[FE] FEAT: keyword style 로직 분리
